+++ lib/can_be/rspec/matchers.rb
@@ -1,0 +1,9 @@
+require 'can_be/rspec/matchers/can_be_detail_matcher'
+require 'can_be/rspec/matchers/can_be_matcher'
+
+module CanBe
+  module RSpec
+    module Matchers
+    end
+  end
+end