+++ spec/spec_helper.rb
@@ -15,6 +15,8 @@ require 'support/models'
 require 'support/can_be_shared_examples'
 require 'support/can_be_detail_shared_examples'
 
+require 'can_be/rspec/matchers'
+
 RSpec.configure do |config|
   config.include ModelMacros